Added project to Atmel Studio 7
Fixed bug related to calibration counter for sensor. Fixed bug regarding steering rad not being returned as a float. Fixed bug regarding PWM degree not beinged signed for output. Fixed bug where timer0 was not triggering interrupt.

+++ src/intellidrive/lib/pid.c
@@ -1,6 +1,6 @@
 #include "pid.h"
 
-uint16_t compute_corrected_yaw(float *ref_yaw, int16_t *yaw_offset, float t)
+float compute_corrected_yaw(float *ref_yaw, int16_t *yaw_offset, float t)
 {
     static float last_error = 0.0;
     static float error_sum = 0.0;
