+++ src/intellidrive/lib/pid.h
@@ -6,6 +6,6 @@
 #include "MPU6050_support.h"
 
 /* PID function */
-uint16_t compute_corrected_yaw(float *ref_yaw, int16_t *yaw_offset, float t);
+float compute_corrected_yaw(float *ref_yaw, int16_t *yaw_offset, float t);
 
 #endif /* PID_H_ */
